requirements.txt
langchain
langchain-openai
openai
sentence-transformers
faiss-cpu
fastapi
uvicorn
python-dotenv
pandas
